XRDDEV-237 add separate module for development dependencies that do not go into production jar
