+++ src/proxy-ui-api-devtools/README.md
@@ -1,0 +1,4 @@
+# Development tools
+
+Module contains development-related classes and configurations.
+These will not be packaged inside the real application jar.
