+++ src/proxy-ui-api-devtools/src/main/resources/application-development.properties
@@ -1,0 +1,5 @@
+# properties for development profile
+
+spring.devtools.remote.secret=development-deploy-secret
+server.ssl.enabled=false
+server.port=8080
