+++ src/proxy-ui-api/src/main/resources/application.properties
@@ -1,4 +1,4 @@
-# properties for dev/prod runtime
+# shared properties for dev/prod runtime
 
 spring.jpa.hibernate.ddl-auto=none
 spring.jpa.hibernate.show-sql=true
@@ -9,6 +9,7 @@ spring.jpa.properties.hibernate.dialect = ee.ria.xroad.common.db.CustomPostgreSQ
 spring.jpa.hibernate.naming.physical-strategy=org.hibernate.boot.model.naming.PhysicalNamingStrategyStandardImpl
 
 # TLS
+# (can be overridden with external ssl.properties)
 server.ssl.key-store=classpath:nginx.p12
 server.ssl.key-store-password=nginx
 server.ssl.enabled=true
