+++ src/proxy-ui-api/src/main/resources/common-application.properties
@@ -5,16 +5,13 @@ proto.pam=true
 #
 # default is 30m
 # server.servlet.session.timeout=1m
-management.endpoints.web.exposure.include=*
-management.endpoint.shutdown.enabled=true
 server.port=8553
-#
-spring.devtools.remote.secret=development-deploy-secret
+
 spring.mvc.throw-exception-if-no-handler-found=true
 spring.resources.add-mappings=false
 #server.error.include-stacktrace=always
 #server.error.include-exception=true
-#
+
 ## *.hbm.xml files are loaded from classpath
 spring.jpa.mapping-resources=identifiers.hbm.xml,serverconf.hbm.xml,apikey.hbm.xml
 spring.jpa.hibernate.ddl-auto=none
