+++ src/settings.gradle
@@ -14,6 +14,7 @@ include "common-op-monitoring"
 include "proxy"
 include "proxy-ui"
 include "proxy-ui-api"
+include "proxy-ui-api-devtools"
 include "signer"
 include "center-ui"
 include "center-util"
